requirements.txt
Flask
Flask-SQLAlchemy
Flask-Admin
Flask-CORS
psycopg2-binary
python-dotenv
